Slate: Move table handlekeydown to new file
