+++ src/components/SlateEditor/plugins/table/utils.ts
@@ -1,16 +1,7 @@
-import { Editor, Element, NodeEntry, Path, Transforms } from 'slate';
+import { Editor, Element, Path, Transforms } from 'slate';
 import { jsx } from 'slate-hyperscript';
 import { ReactEditor } from 'slate-react';
-import {
-  TableCellElement,
-  TableElement,
-  TableRowElement,
-  KEY_ARROW_DOWN,
-  KEY_ARROW_UP,
-  KEY_TAB,
-  TableHeadElement,
-  TableBodyElement,
-} from '.';
+import { TableElement, TableRowElement, TableHeadElement, TableBodyElement } from '.';
 import { defaultParagraphBlock } from '../paragraph/utils';
 import { findCellInMatrix, getTableAsMatrix, getTableBodyAsMatrix } from './matrix';
 
@@ -94,7 +85,7 @@ export const getTableHeight = (element: TableHeadElement | TableBodyElement) => 
   return element.children.length;
 };
 
-const createIdenticalRow = (element: TableRowElement) => {
+export const createIdenticalRow = (element: TableRowElement) => {
   return jsx(
     'element',
     { type: TYPE_TABLE_ROW },
@@ -404,205 +395,4 @@ export const removeTable = (editor: Editor, path: Path) => {
     at: path,
     match: node => Element.isElement(node) && node.type === TYPE_TABLE,
   });
-};
-
-export const handleTableKeydown = (
-  event: KeyboardEvent,
-  editor: Editor,
-  tableEntry: NodeEntry<TableElement>,
-) => {
-  event.preventDefault();
-  if (editor.selection) {
-    const [cellEntry] = Editor.nodes(editor, {
-      at: editor.selection.anchor.path,
-      match: node => Element.isElement(node) && node.type === TYPE_TABLE_CELL,
-    });
-    if (!cellEntry) {
-      return;
-    }
-    const [rowEntry] = Editor.nodes(editor, {
-      at: editor.selection.anchor.path,
-      match: node => Element.isElement(node) && node.type === TYPE_TABLE_ROW,
-    });
-    if (!rowEntry) {
-      return;
-    }
-    const [bodyEntry] = Editor.nodes(editor, {
-      at: editor.selection.anchor.path,
-      match: node =>
-        Element.isElement(node) && (node.type === TYPE_TABLE_HEAD || node.type === TYPE_TABLE_BODY),
-    });
-    if (!bodyEntry) {
-      return;
-    }
-
-    switch (event.key) {
-      case KEY_ARROW_DOWN:
-        return moveDown(
-          editor,
-          tableEntry,
-          rowEntry as NodeEntry<TableRowElement>,
-          cellEntry as NodeEntry<TableCellElement>,
-        );
-      case KEY_ARROW_UP:
-        return moveUp(
-          editor,
-          tableEntry,
-          rowEntry as NodeEntry<TableRowElement>,
-          cellEntry as NodeEntry<TableCellElement>,
-        );
-      case KEY_TAB:
-        if (event.shiftKey) {
-          return moveLeft(
-            editor,
-            tableEntry,
-            bodyEntry as NodeEntry<TableHeadElement | TableBodyElement>,
-            rowEntry as NodeEntry<TableRowElement>,
-            cellEntry as NodeEntry<TableCellElement>,
-          );
-        }
-        return moveRight(
-          editor,
-          tableEntry,
-          bodyEntry as NodeEntry<TableHeadElement | TableBodyElement>,
-          rowEntry as NodeEntry<TableRowElement>,
-          cellEntry as NodeEntry<TableCellElement>,
-        );
-
-      default:
-        return;
-    }
-  }
-};
-
-const moveLeft = (
-  editor: Editor,
-  tableEntry: NodeEntry<TableElement>,
-  bodyEntry: NodeEntry<TableBodyElement | TableHeadElement>,
-  rowEntry: NodeEntry<TableRowElement>,
-  cellEntry: NodeEntry<TableCellElement>,
-) => {
-  const tablePath = tableEntry[1];
-  const bodyPath = bodyEntry[1];
-  const [row, rowPath] = rowEntry;
-  const cellPath = cellEntry[1];
-  if (
-    (Path.hasPrevious(cellPath) || Path.hasPrevious(rowPath) || Path.hasPrevious(bodyPath)) &&
-    editor.selection
-  ) {
-    Transforms.select(editor, Editor.start(editor, cellEntry[1]));
-    Transforms.move(editor, { reverse: true });
-    Transforms.select(editor, Editor.range(editor, editor.selection.anchor.path));
-    return;
-  }
-
-  if (Path.equals([...tablePath, 0, 0, 0], cellPath)) {
-    const targetPath = [...tablePath, 0, 0];
-    Transforms.insertNodes(editor, createIdenticalRow(row), { at: targetPath });
-    Transforms.select(editor, {
-      anchor: Editor.point(editor, targetPath, { edge: 'start' }),
-      focus: Editor.point(editor, targetPath, { edge: 'start' }),
-    });
-  }
-};
-
-const moveRight = (
-  editor: Editor,
-  tableEntry: NodeEntry<TableElement>,
-  bodyEntry: NodeEntry<TableBodyElement | TableHeadElement>,
-  rowEntry: NodeEntry<TableRowElement>,
-  cellEntry: NodeEntry<TableCellElement>,
-) => {
-  const tablePath = tableEntry[1];
-  const bodyPath = bodyEntry[1];
-  const [row, rowPath] = rowEntry;
-  const cellPath = cellEntry[1];
-  const nextPath = Path.next(cellPath);
-  const nextRowPath = Path.next(rowPath);
-  const nextBodyPath = Path.next(bodyPath);
-
-  if (
-    (Editor.hasPath(editor, nextPath) ||
-      Editor.hasPath(editor, nextRowPath) ||
-      Editor.hasPath(editor, nextBodyPath)) &&
-    editor.selection
-  ) {
-    Transforms.select(editor, Editor.end(editor, cellEntry[1]));
-    Transforms.move(editor);
-    Transforms.select(editor, Editor.range(editor, editor.selection.anchor.path));
-    return;
-  }
-
-  const TableEndPoint = Editor.point(editor, tablePath, { edge: 'end' });
-  if (Path.isDescendant(TableEndPoint.path, cellPath)) {
-    const targetPath = [...bodyPath, row.children.length];
-    Transforms.insertNodes(editor, defaultTableRowBlock(1), { at: targetPath });
-    Transforms.select(editor, {
-      anchor: Editor.point(editor, tablePath, { edge: 'end' }),
-      focus: Editor.point(editor, tablePath, { edge: 'end' }),
-    });
-  }
-};
-
-const moveDown = (
-  editor: Editor,
-  tableEntry: NodeEntry<TableElement>,
-  rowEntry: NodeEntry<TableRowElement>,
-  cellEntry: NodeEntry<TableCellElement>,
-) => {
-  const tablePath = tableEntry[1];
-
-  const rowPath = rowEntry[1];
-
-  const cellPath = cellEntry[1];
-  const columnIndex = cellPath[cellPath.length - 1];
-
-  const nextPath = [...Path.next(rowPath), columnIndex];
-
-  if (Editor.hasPath(editor, nextPath)) {
-    Transforms.select(editor, {
-      anchor: Editor.point(editor, nextPath, { edge: 'start' }),
-      focus: Editor.point(editor, nextPath, { edge: 'start' }),
-    });
-  } else if (
-    Editor.hasPath(editor, Path.next(tablePath)) &&
-    !Editor.hasPath(editor, Path.next(rowPath))
-  ) {
-    Transforms.select(editor, {
-      anchor: Editor.point(editor, Path.next(tablePath), { edge: 'start' }),
-      focus: Editor.point(editor, Path.next(tablePath), { edge: 'start' }),
-    });
-  }
-};
-
-const moveUp = (
-  editor: Editor,
-  tableEntry: NodeEntry<TableElement>,
-  rowEntry: NodeEntry<TableRowElement>,
-  cellEntry: NodeEntry<TableCellElement>,
-) => {
-  const tablePath = tableEntry[1];
-
-  const rowPath = rowEntry[1];
-
-  const cellPath = cellEntry[1];
-  const columnIndex = cellPath[cellPath.length - 1];
-
-  if (Path.hasPrevious(rowPath)) {
-    const previousPath = [...Path.previous(rowPath), columnIndex];
-    if (Editor.hasPath(editor, previousPath)) {
-      Transforms.select(editor, {
-        anchor: Editor.point(editor, previousPath, { edge: 'start' }),
-        focus: Editor.point(editor, previousPath, { edge: 'start' }),
-      });
-    }
-  } else if (Path.hasPrevious(tablePath)) {
-    const previousPath = Path.previous(tablePath);
-    if (Editor.hasPath(editor, previousPath)) {
-      Transforms.select(editor, {
-        anchor: Editor.point(editor, previousPath, { edge: 'start' }),
-        focus: Editor.point(editor, previousPath, { edge: 'start' }),
-      });
-    }
-  }
 };
